relocate util methods
